# openapi.json
{"openapi":"3.1.0","info":{"title":"Image Analysis API","description":"High-performance image analysis API service","version":"1.0.0"},"paths":{"/health/":{"get":{"tags":["health"],"summary":"Health Check","description":"Проверка здоровья сервиса и его зависимостей.\n\nReturns:\n    Статус здоровья сервиса","operationId":"health_check_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/ready":{"get":{"tags":["health"],"summary":"Readiness Check","description":"Проверка готовности сервиса принимать трафик.\n\nReturns:\n    Статус готовности","operationId":"readiness_check_health_ready_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/health/live":{"get":{"tags":["health"],"summary":"Liveness Check","description":"Liveness probe для Kubernetes.\n\nReturns:\n    Статус alive","operationId":"liveness_check_health_live_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/analyze":{"post":{"tags":["analysis"],"summary":"Анализ документа","description":"Загружает и анализирует документ (PDF или изображение) на наличие лиц","operationId":"analyze_document_analyze_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_analyze_document_analyze_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TaskStatusResponse"}}}},"400":{"description":"Некорректный запрос","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"401":{"description":"Не авторизован","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"413":{"description":"Размер файла превышает 10MB","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"422":{"description":"PDF не содержит изображений","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"429":{"description":"Превышен лимит запросов","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}},"500":{"description":"Внутренняя ошибка сервера","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorResponse"}}}}}}},"/metrics":{"get":{"tags":["metrics"],"summary":"Метрики Prometheus","description":"Возвращает метрики в формате Prometheus","operationId":"get_metrics_metrics_get","responses":{"200":{"description":"Successful Response","content":{"text/plain":{"schema":{"type":"string"}}}}}}},"/auth/api-keys":{"get":{"tags":["auth"],"summary":"List Api Keys","description":"Получить список API-ключей пользователя.\n\nReturns:\n    Список API-ключей без полного ключа (только метаданные)","operationId":"list_api_keys_auth_api_keys_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/APIKeyResponse"},"type":"array","title":"Response List Api Keys Auth Api Keys Get"}}}}}},"post":{"tags":["auth"],"summary":"Create Api Key","description":"Создать новый API-ключ.\n\nReturns:\n    API-ключ с полным значением ключа (возвращается только один раз)","operationId":"create_api_key_auth_api_keys_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAPIKeyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/APIKeyResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/api-keys/{key_id}/rotate":{"post":{"tags":["auth"],"summary":"Rotate Api Key","description":"Ротировать существующий API-ключ.\n\nArgs:\n    key_id: ID API-ключа для ротации\n\nReturns:\n    Новый API-ключ и сообщение об успехе","operationId":"rotate_api_key_auth_api_keys__key_id__rotate_post","parameters":[{"name":"key_id","in":"path","required":true,"schema":{"type":"integer","title":"Key Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"},"title":"Response Rotate Api Key Auth Api Keys  Key Id  Rotate Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/auth/api-keys/{key_id}":{"delete":{"tags":["auth"],"summary":"Delete Api Key","description":"Деактивировать API-ключ.\n\nArgs:\n    key_id: ID API-ключа для деактивации\n\nReturns:\n    Статус успеха и сообщение","operationId":"delete_api_key_auth_api_keys__key_id__delete","parameters":[{"name":"key_id","in":"path","required":true,"schema":{"type":"integer","title":"Key Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"boolean"},"title":"Response Delete Api Key Auth Api Keys  Key Id  Delete"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/telegram/telegram":{"post":{"tags":["auth"],"summary":"Telegram Auth","description":"Аутентификация через Telegram WebApp.\n\nПроцесс:\n1. Получает init_data от клиента\n2. Верифицирует подпись через HMAC-SHA256\n3. Проверяет, что auth_date не старше 24 часов\n4. Находит или создает пользователя по telegram_id\n5. Создает JWT токены\n6. Создает API-ключ для использования\n7. Возвращает токены и API-ключ\n\nReturns:\n    TelegramAuthResponse с токенами и данными пользователя\n\nRaises:\n    HTTPException 401: Неверная подпись или просроченные данные\n    HTTPException 500: Ошибка сервера","operationId":"telegram_auth_telegram_telegram_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramAuthRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TelegramAuthResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/telegram/refresh":{"post":{"tags":["auth"],"summary":"Refresh Token","description":"Обновление access токена по refresh токену.\n\nArgs:\n    refresh_token: Refresh токен\n\nReturns:\n    TokenResponse с новыми токенами\n\nRaises:\n    HTTPException 401: Неверный refresh токен\n    HTTPException 500: Ошибка сервера","operationId":"refresh_token_telegram_refresh_post","parameters":[{"name":"refresh_token","in":"query","required":true,"schema":{"type":"string","title":"Refresh Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Root","description":"Корневой эндпоинт с информацией о сервисе.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"APIKeyResponse":{"properties":{"id":{"type":"integer","title":"Id"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key"},"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name"},"is_active":{"type":"boolean","title":"Is Active"},"expires_at":{"anyOf":[{"type":"string","format":"date-time"},{"type":"null"}],"title":"Expires At"},"created_at":{"type":"string","format":"date-time","title":"Created At"}},"type":"object","required":["id","name","is_active","expires_at","created_at"],"title":"APIKeyResponse","description":"Ответ с данными API ключа."},"Body_analyze_document_analyze_post":{"properties":{"file":{"type":"string","contentMediaType":"application/octet-stream","title":"File","description":"Документ для анализа (PDF, JPEG, PNG, WebP или BMP)"}},"type":"object","required":["file"],"title":"Body_analyze_document_analyze_post"},"CreateAPIKeyRequest":{"properties":{"name":{"anyOf":[{"type":"string","maxLength":100},{"type":"null"}],"title":"Name"},"expires_days":{"anyOf":[{"type":"integer","maximum":365.0,"minimum":1.0},{"type":"null"}],"title":"Expires Days"}},"type":"object","title":"CreateAPIKeyRequest","description":"Запрос создания API ключа."},"ErrorResponse":{"properties":{"error":{"type":"string","title":"Error","description":"Тип ошибки"},"message":{"type":"string","title":"Message","description":"Сообщение об ошибке"},"details":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Details","description":"Дополнительные детали"},"request_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Request Id","description":"ID запроса"},"timestamp":{"type":"string","title":"Timestamp","description":"Время ошибки в ISO формате"}},"type":"object","required":["error","message","timestamp"],"title":"ErrorResponse","description":"Стандартизированный ответ об ошибке.\n\nAttributes:\n    error: Тип ошибки\n    message: Человекочитаемое сообщение\n    details: Дополнительные детали ошибки (опционально)\n    request_id: ID запроса для трассировки\n    timestamp: Время возникновения ошибки","example":{"details":{"field":"image","issue":"file too large"},"error":"validation_error","message":"Неверные данные в запросе","request_id":"req_123456789","timestamp":"2024-01-15T10:30:00Z"}},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"TaskStatus":{"type":"string","enum":["pending","processing","completed","failed"],"title":"TaskStatus","description":"Статус задачи анализа."},"TaskStatusResponse":{"properties":{"task_id":{"type":"string","title":"Task Id","description":"ID задачи"},"status":{"$ref":"#/components/schemas/TaskStatus","description":"Статус задачи"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result","description":"Результат анализа"},"message":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Message","description":"Сообщение о статусе"}},"type":"object","required":["task_id","status"],"title":"TaskStatusResponse","description":"Ответ о статусе задачи.\n\nAttributes:\n    task_id: ID задачи\n    status: Статус задачи\n    result: Результат (если завершена)\n    message: Сообщение о статусе","example":{"message":"Анализ завершен","result":{"faces":[],"filename":"document.pdf","summary":"Found 2 image(s), 3 face(s).","total_faces":3,"total_images":2},"status":"completed","task_id":"550e8400-e29b-41d4-a716-446655440000"}},"TelegramAuthRequest":{"properties":{"init_data":{"type":"string","title":"Init Data","description":"Строка query параметров от Telegram"}},"type":"object","required":["init_data"],"title":"TelegramAuthRequest","description":"Запрос аутентификации через Telegram."},"TelegramAuthResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"},"user":{"additionalProperties":true,"type":"object","title":"User"},"api_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Api Key","description":"API ключ для использования"}},"type":"object","required":["access_token","refresh_token","user"],"title":"TelegramAuthResponse","description":"Ответ при успешной аутентификации через Telegram."},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"refresh_token":{"type":"string","title":"Refresh Token"},"token_type":{"type":"string","title":"Token Type","default":"bearer"}},"type":"object","required":["access_token","refresh_token"],"title":"TokenResponse","description":"Ответ с JWT токенами."},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}},"servers":[{"url":"/api/v1"}]}